backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.0.2
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27
pandas==2.1.4
openpyxl==3.1.2
PyPDF2==3.0.1
markdown==3.5.1
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1 
